# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Food Ordering Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl: string | null;
  createdAt: string;
}

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
}

export interface CategoryUpdate {
  name?: string;
  slug?: string;
  description?: string;
  imageUrl?: string;
}

export interface MenuItem {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  available: boolean;
  featured: boolean;
  /** @nullable */
  prepTimeMinutes?: number | null;
  createdAt: string;
}

export interface MenuItemInput {
  name: string;
  description?: string;
  price: number;
  imageUrl?: string;
  categoryId: number;
  available?: boolean;
  featured?: boolean;
  prepTimeMinutes?: number;
}

export interface MenuItemUpdate {
  name?: string;
  description?: string;
  price?: number;
  imageUrl?: string;
  categoryId?: number;
  available?: boolean;
  featured?: boolean;
  prepTimeMinutes?: number;
}

export interface OrderItem {
  menuItemId: number;
  name: string;
  quantity: number;
  unitPrice: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  preparing: 'preparing',
  ready: 'ready',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type OrderPaymentMethod = typeof OrderPaymentMethod[keyof typeof OrderPaymentMethod];


export const OrderPaymentMethod = {
  cash_on_delivery: 'cash_on_delivery',
  jazz_cash: 'jazz_cash',
  easy_paisa: 'easy_paisa',
} as const;

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerAddress?: string | null;
  status: OrderStatus;
  total: number;
  /** @nullable */
  notes?: string | null;
  paymentMethod: OrderPaymentMethod;
  items: OrderItem[];
  createdAt: string;
}

export type OrderInputPaymentMethod = typeof OrderInputPaymentMethod[keyof typeof OrderInputPaymentMethod];


export const OrderInputPaymentMethod = {
  cash_on_delivery: 'cash_on_delivery',
  jazz_cash: 'jazz_cash',
  easy_paisa: 'easy_paisa',
} as const;

export type OrderInputItemsItem = {
  menuItemId: number;
  quantity: number;
};

export interface OrderInput {
  customerName: string;
  customerPhone: string;
  customerAddress?: string;
  notes?: string;
  paymentMethod: OrderInputPaymentMethod;
  items: OrderInputItemsItem[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  preparing: 'preparing',
  ready: 'ready',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface DashboardStats {
  totalOrders: number;
  totalRevenue: number;
  pendingOrders: number;
  totalItems: number;
  totalCategories: number;
  todayOrders: number;
  todayRevenue: number;
}

export interface TopItem {
  id: number;
  name: string;
  orderCount: number;
  revenue: number;
  /** @nullable */
  imageUrl: string | null;
}

export interface Rider {
  id: number;
  name: string;
  phone: string;
  active: boolean;
  /** @nullable */
  restaurantId?: number | null;
}

export interface RiderInput {
  name?: string;
  phone?: string;
  active?: boolean;
}

export type OrderPaymentStatusUpdatePaymentStatus = typeof OrderPaymentStatusUpdatePaymentStatus[keyof typeof OrderPaymentStatusUpdatePaymentStatus];


export const OrderPaymentStatusUpdatePaymentStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  not_required: 'not_required',
} as const;

export interface OrderPaymentStatusUpdate {
  paymentStatus: OrderPaymentStatusUpdatePaymentStatus;
}

export interface RiderAssignment {
  /** @nullable */
  riderId?: number | null;
}

export interface UploadUrlRequest {
  /**
     * Original file name.
     * @minLength 1
     */
  name: string;
  /**
     * File size in bytes.
     * @minimum 1
     */
  size: number;
  /**
     * MIME type of the file (e.g. `image/jpeg`).
     * @minLength 1
     */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database. */
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export type ListItemsParams = {
categoryId?: number;
featured?: boolean;
};

export type ListOrdersParams = {
status?: ListOrdersStatus;
dateFrom?: string;
dateTo?: string;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  preparing: 'preparing',
  ready: 'ready',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type DeleteRider200 = {
  success?: boolean;
};